PRETEXTA/backend/requirements.txt
# Core FastAPI and web framework
fastapi
uvicorn
python-dotenv
pydantic

# Database
motor  # AsyncIO MongoDB driver

# Authentication & Security
bcrypt
PyJWT

# LangChain ecosystem
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-groq

# Utilities
PyYAML
httpx

# Development
ruff
